+++ doc/CHANGELOG.md
@@ -17,6 +17,7 @@ software or firmware builds or that the documentation is consistent.
 The full set of tests is only run for tagged releases._
 
 ### Summary
+- 23 line interrupt mapper now for full system configuration
 - man pages now [available online](http://www.retro11.de/manp/w11/man/cat1/).
 - add `sysmon_rbus` in `sys_tst_sram_n4`
 - Auto-dection of Digilent boards with `FT2232HQ` interface for